+++ eslint.config.js
@@ -1,4 +1,5 @@
 import eslint from "@eslint/js";
+import importPlugin from "eslint-plugin-import";
 import stylistic from "@stylistic/eslint-plugin";
 import reactPlugin from "eslint-plugin-react";
 import globals from "globals";
@@ -15,10 +16,23 @@ export default tseslint.config(
       "dist",
       "eslint.config.js",
       "node_modules",
-      "src/components/ThemeProvider/*",
+      "src/app/Provider.tsx",
       "src/components/ui/*",
       "src/lib/utils.ts",
     ]
+  },
+  {
+    files: ["**/*.{ts,tsx}"],
+    extends: [
+      importPlugin.flatConfigs.recommended,
+      importPlugin.flatConfigs.typescript
+    ],
+    settings: {
+      "import/resolver": {
+        "typescript": true,
+        "node": true,
+      },
+    },
   },
   {
     languageOptions: {
@@ -46,11 +60,45 @@ export default tseslint.config(
   {
     rules: {
       "@stylistic/indent": ["error", 2],
+      "@stylistic/quotes": ["error", "double"],
       "@stylistic/semi": "error",
       "@typescript-eslint/consistent-type-definitions": "off",
       "@typescript-eslint/explicit-function-return-type": "off",
       "@typescript-eslint/no-non-null-assertion": "off",
       "@typescript-eslint/no-unnecessary-boolean-literal-compare": "off",
+      "import/no-cycle": "error",
+      "import/no-restricted-paths": ["error", {
+        "zones": [
+          {
+            target: "./src/features/Auth",
+            from: "./src/features",
+            except: ["./Auth"],
+          },
+          {
+            target: "./src/features/Favorites",
+            from: "./src/features",
+            except: ["./Favorites"],
+          },
+          {
+            target: "./src/features/Search",
+            from: "./src/features",
+            except: ["./Search"],
+          },
+          {
+            target: "./src/features",
+            from: "./src/app",
+          },
+          {
+            target: [
+              "./src/components",
+              "./src/lib",
+              "./src/types",
+              "./src/utils",
+            ],
+            from: ["./src/app", "./src/features"],
+          },
+        ]
+      }],
       "react/react-in-jsx-scope": "off",
     },
   },
