+++ package.json
@@ -19,7 +19,7 @@
     "class-variance-authority": "0.7.1",
     "clsx": "2.1.1",
     "immer": "^10.1.1",
-    "lucide-react": "0.485.0",
+    "lucide-react": "0.486.0",
     "next-themes": "0.4.6",
     "react": "18.3.1",
     "react-dom": "18.3.1",
@@ -41,11 +41,13 @@
     "@types/react-dom": "19.0.4",
     "@vitejs/plugin-react": "4.3.4",
     "eslint": "9.23.0",
+    "eslint-import-resolver-typescript": "^4.3.1",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-react": "7.37.4",
     "globals": "15.15.0",
     "typescript": "5.8.2",
-    "typescript-eslint": "8.28.0",
-    "vite": "6.2.3"
+    "typescript-eslint": "8.29.0",
+    "vite": "6.2.4"
   },
   "engines": {
     "node": "22",
