--- src/components/ThemeProvider/ThemeProvider.tsx
@@ -1,58 +1,0 @@
-import { createContext, useContext, useEffect, useState } from "react";
-
-type ThemeProviderProps = {
-  children: React.ReactNode;
-  defaultTheme?: "dark";
-  storageKey?: string;
-}
-
-type ThemeProviderState = {
-  theme: "dark";
-  setTheme: (theme: "dark") => void;
-}
-
-const initialState: ThemeProviderState = {
-  theme: "dark",
-  setTheme: () => null
-}
-
-const ThemeProviderContext = createContext<ThemeProviderState>(initialState);
-
-export function ThemeProvider({
-  children,
-  defaultTheme = "dark",
-  storageKey = "vite-ui-theme",
-  ...props
-}: ThemeProviderProps) {
-  const [theme, setTheme] = useState<"dark">(
-    () => (localStorage.getItem(storageKey) as "dark") || defaultTheme
-  );
-
-  useEffect(() => {
-    const root = window.document.documentElement;
-    root.classList.remove("light", "dark");
-    root.classList.add(theme);
-  }, [theme]);
-
-  const value = {
-    theme,
-    setTheme: (theme: "dark") => {
-      localStorage.setItem(storageKey, theme);
-      setTheme(theme);
-    },
-  };
-
-  return (
-    <ThemeProviderContext.Provider {...props} value={value}>
-      {children}
-    </ThemeProviderContext.Provider>
-  );
-}
-
-export const useTheme = () => {
-  const context = useContext(ThemeProviderContext);
-  if (context === undefined) {
-    throw new Error("useTheme must be used within a ThemeProvider");
-  }
-  return context;
-}
