+++ src/features/Favorites/stores/FavoritesStore.ts
@@ -1,0 +1,42 @@
+import type {
+  FavoritesActions,
+  FavoritesState
+} from "@/features/Favorites/types/Favorites";
+import type { Dog } from "@/types/Dog";
+import { create } from "zustand";
+import { devtools, persist } from "zustand/middleware";
+import { immer } from "zustand/middleware/immer";
+
+const initialState: FavoritesState = {
+  favoriteDogs: {},
+};
+
+const useFavoritesStore = create<FavoritesState & FavoritesActions>()(
+  devtools(
+    persist(
+      immer((set) => ({
+        ...initialState,
+        addFavoriteDog: (dog: Dog) => {
+          set((state) => {
+            state.favoriteDogs[dog.id] = dog;
+          });
+        },
+        removeFavoriteDog: (dogId: string) => {
+          set((state) => {
+            const map = { ...state.favoriteDogs };
+            delete map[dogId];  // eslint-disable-line
+            return { favoriteDogs: map };
+          });
+        },
+        resetFavoriteDogs: () => {
+          set({ ...initialState });
+        },
+      })),
+      { name: "pupperfield-dog-state" },
+    )
+  ),
+);
+
+export {
+  useFavoritesStore as default
+};
