+++ src/types/Auth.ts
@@ -1,9 +1,9 @@
-export type LogInRequestBody = {
-  email: string;
-  name: string;
-};
-
 export type AuthResponse = {
   error?: unknown;
   ok: boolean;
+};
+
+export type LogInRequestBody = {
+  email: string;
+  name: string;
 };
