+++ tsconfig.app.json
@@ -28,7 +28,7 @@
       "@/*": ["./src/*"],
       "@assets/*": ["./src/assets/*"],
       "@components/*": ["./src/components/*"],
-      "@pages/*": ["./src/pages/*"],
+      "@pages/*": ["src/features/*"],
       "@services/*": ["./src/services/*"],
       "@slices/*": ["src/slices/*"],
     }
