+++ vite.config.ts
@@ -1,7 +1,7 @@
 import tailwindcss from "@tailwindcss/vite";
-import react from '@vitejs/plugin-react';
+import react from "@vitejs/plugin-react";
 import path from "node:path";
-import { defineConfig } from 'vite';
+import { defineConfig } from "vite";
 
 // https://vite.dev/config/
 export default defineConfig({
@@ -10,13 +10,6 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
-      "@assets": path.resolve(__dirname, "./src/assets"),
-      "@components": path.resolve(__dirname, "./src/components"),
-      "@lib": path.resolve(__dirname, "./src/lib"),
-      "@pages": path.resolve(__dirname, "./src/pages"),
-      "@services": path.resolve(__dirname, "./src/services"),
-      "@stores": path.resolve(__dirname, "./src/stores"),
-      "@types": path.resolve(__dirname, "./src/types"),
     },
   },
 });
